:hammer: chore: re-use code for taxi & tow
